+++ pkg/storage/gc/gc_test.go
@@ -8,19 +8,20 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package engine
+package gc
 
 import (
 	"testing"
 
 	"github.com/cockroachdb/cockroach/pkg/config"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/storage/engine"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 )
 
-func mvccVersionKey(key roachpb.Key, ts hlc.Timestamp) MVCCKey {
-	return MVCCKey{Key: key, Timestamp: ts}
+func mvccVersionKey(key roachpb.Key, ts hlc.Timestamp) engine.MVCCKey {
+	return engine.MVCCKey{Key: key, Timestamp: ts}
 }
 
 var (
@@ -48,7 +49,7 @@ func TestGarbageCollectorFilter(t *testing.T) {
 	testData := []struct {
 		gc       GarbageCollector
 		time     hlc.Timestamp
-		keys     []MVCCKey
+		keys     []engine.MVCCKey
 		values   [][]byte
 		expIdx   int
 		expDelTS hlc.Timestamp
